Extract Here API specific code from WebTileDataSource
 *  github issue #1649
